# docs/swagger.yaml
basePath: /
definitions:
...
  request.CreateUserRequest:
    properties:
      password:
        type: string
      role:
        type: integer
      username:
        type: string
    type: object
...
  request.UpdateUserRequest:
    properties:
      is_activated:
        type: integer
      role:
        type: integer
      username:
        type: string
    type: object
  response.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: API yang digunakan untuk website SIPEKOM
  title: API Sipekom
  version: 0.0.1
paths:
...
  /api/user/:
    get:
      description: get all User
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "get all User [guest\U0001F512]."
      tags:
      - User
  /api/user/create:
    post:
      consumes:
      - application/json
      description: create new User.
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.CreateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "create User [konsulen\U0001F512, mahasiswa\U0001F512, guest\U0001F512]."
      tags:
      - User
  /api/user/data/{id_user}:
    get:
      description: get User data by id based on their role
      parameters:
      - description: User ID
        in: path
        name: id_user
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "get User data based on role [guest\U0001F512]."
      tags:
      - User
  /api/user/delete/{id_user}:
    delete:
      description: delete User by id, mahasiswa and konsulen only can delete their
        own account
      parameters:
      - description: User ID
        in: path
        name: id_user
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "delete User [konsulen \U0001F9F1, mahasiswa \U0001F9F1, guest\U0001F512]."
      tags:
      - User
  /api/user/get/{id_user}:
    get:
      description: get User by id.
      parameters:
      - description: User ID
        in: path
        name: id_user
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "get User [guest\U0001F512]."
      tags:
      - User
  /api/user/update/{id_user}:
    put:
      description: update User by id, mahasiswa and konsulen only can update their
        own account not account data refer to specific role
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateUserRequest'
      - description: User ID
        in: path
        name: id_user
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "update User [konsulen \U0001F9F1, mahasiswa \U0001F9F1, guest\U0001F512]."
      tags:
      - User
...
securityDefinitions:
  ApiKeyAuth:
    description: API Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"